feat(FX-4603): Activity Panel Editorial prefix should be grey (#11902)
