+++ src/Components/Notifications/NotificationItem.tsx
@@ -10,7 +10,11 @@ import { useTracking } from "react-tracking"
 import { useSystemContext } from "System"
 import createLogger from "Utils/logger"
 import { markNotificationAsRead } from "Components/Notifications/Mutations/markNotificationAsRead"
-import { isArtworksBasedNotification } from "./util"
+import {
+  isArtworksBasedNotification,
+  shouldDisplayNotificationTypeLabel,
+} from "./util"
+import { NotificationTypeLabel } from "./NotificationTypeLabel"
 
 const logger = createLogger("NotificationItem")
 
@@ -28,18 +32,6 @@ const NotificationItem: React.FC<NotificationItemProps> = ({ item }) => {
   const shouldDisplayCounts =
     isArtworksBasedNotification(item.notificationType) &&
     remainingArtworksCount > 0
-
-  const getNotificationType = () => {
-    if (item.notificationType === "ARTWORK_ALERT") {
-      return "Alert"
-    }
-    if (item.notificationType === "ARTICLE_FEATURED_ARTIST") {
-      return "Artsy Editorial"
-    }
-
-    return null
-  }
-  const notificationTypeLabel = getNotificationType()
 
   const markAsRead = async () => {
     if (!relayEnvironment) {
@@ -76,12 +68,8 @@ const NotificationItem: React.FC<NotificationItemProps> = ({ item }) => {
     <NotificationItemLink to={item.targetHref} onClick={handlePress}>
       <Flex flex={1} flexDirection="column">
         <Text variant="xs" color="black60">
-          {notificationTypeLabel && (
-            <NotificationTypeLabel
-              aria-label={`Notification type: ${notificationTypeLabel}`}
-            >
-              {notificationTypeLabel} •{" "}
-            </NotificationTypeLabel>
+          {shouldDisplayNotificationTypeLabel(item.notificationType) && (
+            <NotificationTypeLabel notificationType={item.notificationType} />
           )}
           {item.publishedAt}
         </Text>
@@ -172,10 +160,6 @@ export const NotificationItemFragmentContainer = createFragmentContainer(
     `,
   }
 )
-
-const NotificationTypeLabel = styled.span`
-  color: ${themeGet("colors.blue100")};
-`
 
 const NotificationItemLink = styled(RouterLink)`
   display: flex;
