+++ src/Components/Notifications/util.ts
@@ -14,3 +14,9 @@ export const isArtworksBasedNotification = (
 ) => {
   return ["ARTWORK_ALERT", "ARTWORK_PUBLISHED"].includes(notificationType)
 }
+
+export const shouldDisplayNotificationTypeLabel = (
+  notificationType: string
+) => {
+  return ["ARTWORK_ALERT", "ARTICLE_FEATURED_ARTIST"].includes(notificationType)
+}
